MMFMWK-9336 imxvideoconvert: fix green strip when switching between streams
When dynamically switching between streams such as one 1080p H264
and one 1080p HEVC without destroying the pipeline and restart,
a strip appear on the bottom of the second video on 8MP.

This is because decoded H264 video buffer size is 1920x1088 which has
8 padding at bottom, while decoded HEVC video buffer size is 1920x1080
which doesn't need padding. First video's 8 padding is added to the
second video's bottom. It causes to set wrong value for second video's
g2d->src.base.height. Need to correct it using actual video meta alignment.

./single_switch -s -l -i -t 10 file:///run/media/KINGSTON-sda1/bbb_h264_1920x1080_30fps_yuv420p_highprofile_level4p2_10Mbps_60s.mp4  file:///run/media/KINGSTON-sda1/btox_1080p_10_mbps_h265.mp4

Signed-off-by: Hou Qi <qi.hou@nxp.com>