LF-14096 aiurdemux: fix nonseekable http stream cannot start play issue
For nonseekable http stream, aiurdemux pass LIVE flags to parser when
create core parser. Parser will return a huge TS for the first frame.
When. aiurdemux will adjust timestamp just like live stream in start_offset()
but the pipeline is not live, when aiurdemux get pipeline clock, it is
not ready yet because pipeline only set clock when PAUSE->PLAYING,
aiurdemux will leave the huge TS and push it down, which cause pipeline hang
on sink.

Solution is take the first TS as base and disable start_offset in case
it adjust ts after it got valid pipeline clock.

Signed-off-by: Haihua Hu <jared.hu@nxp.com>